Move control methods to flip card state
